--- simaas/tests/e2e/__init__.py
@@ -1,1 +1,0 @@
-# End-to-end tests package
