--- simaas/tests/integration/__init__.py
@@ -1,1 +1,0 @@
-# Integration tests package
