--- simaas/tests/integration/processors/__init__.py
@@ -1,1 +1,0 @@
-# Processor integration tests package
