--- simaas/tests/integration/services/__init__.py
@@ -1,1 +1,0 @@
-# Service integration tests package
