+++ simaas/tests/test_cli_identity.py
@@ -1,0 +1,403 @@
+import asyncio
+import json
+import logging
+import os
+import shutil
+import socket
+import tempfile
+import threading
+import time
+import traceback
+from typing import List, Union, Any, Optional, Tuple
+
+import pytest
+from docker.errors import ImageNotFound
+from git import Repo
+
+from simaas.cli.cmd_image import clone_repository, build_processor_image, PDIBuildLocal, PDIBuildGithub, PDIExport, \
+    PDIImport, PDIMetaInformation
+from simaas.cli.cmd_namespace import NamespaceUpdate, NamespaceList, NamespaceShow
+from simaas.cli.cmd_service import Service
+
+from examples.simple.abc.processor import write_value
+from simaas.core.helpers import get_timestamp_now
+
+from simaas.nodedb.protocol import P2PJoinNetwork, P2PLeaveNetwork
+from simaas.nodedb.schemas import NamespaceInfo
+from simaas.rti.base import DBJobInfo
+from plugins.rti_docker import DefaultRTIService
+from simaas.cli.cmd_dor import DORAdd, DORMeta, DORDownload, DORRemove, DORSearch, DORTag, DORUntag, DORAccessShow, \
+    DORAccessGrant, DORAccessRevoke
+from simaas.cli.cmd_identity import IdentityCreate, IdentityList, IdentityRemove, IdentityShow, IdentityDiscover, \
+    IdentityPublish, IdentityUpdate, CredentialsList, CredentialsAddGithubCredentials, CredentialsRemove
+from simaas.cli.cmd_job_runner import JobRunner
+from simaas.cli.cmd_network import NetworkList
+from simaas.cli.cmd_rti import RTIProcDeploy, RTIProcList, RTIProcShow, RTIProcUndeploy, RTIJobSubmit, RTIJobStatus, \
+    RTIJobList, RTIJobCancel, RTIVolumeCreateFSRef, RTIVolumeList, RTIVolumeDelete, RTIVolumeCreateEFSRef
+from simaas.cli.exceptions import CLIRuntimeError
+from simaas.core.identity import Identity
+from simaas.core.keystore import Keystore
+from simaas.core.logging import Logging
+from simaas.dor.api import DORProxy
+from simaas.dor.schemas import DataObject, ProcessorDescriptor, GitProcessorPointer
+from simaas.helpers import find_available_port, docker_export_image, PortMaster, determine_local_ip, find_processors
+from simaas.node.base import Node
+from simaas.node.default import DefaultNode
+from simaas.p2p.base import P2PAddress
+from simaas.rti.protocol import P2PInterruptJob
+from simaas.rti.schemas import Task, Job, JobStatus, Severity, ExitCode, JobResult, Processor
+from simaas.core.processor import ProgressListener, ProcessorBase, ProcessorRuntimeError, Namespace
+from simaas.tests.conftest import REPOSITORY_COMMIT_ID, REPOSITORY_URL, PROC_ABC_PATH
+
+logger = Logging.get(__name__)
+repo_root_path = os.path.abspath(os.path.join(os.path.dirname(os.path.abspath(__file__)), '..', '..'))
+examples_path = os.path.join(repo_root_path, 'examples')
+
+
+@pytest.fixture(scope="session")
+def temp_dir():
+    with tempfile.TemporaryDirectory() as tempdir:
+        yield tempdir
+
+
+
+def test_cli_identity_crud(temp_dir):
+    """Test CLI identity management workflow (list, create, show, remove)."""
+    # list all identities
+    try:
+        args = {
+            'keystore': temp_dir,
+        }
+
+        cmd = IdentityList()
+        result = cmd.execute(args)
+        assert result is not None
+        assert 'available' in result
+        assert len(result['available']) == 0
+
+    except CLIRuntimeError:
+        assert False
+
+    # create an identity
+    try:
+        args = {
+            'keystore': temp_dir,
+            'name': 'name',
+            'email': 'email',
+            'password': 'password'
+        }
+
+        cmd = IdentityCreate()
+        result = cmd.execute(args)
+        assert result is not None
+        assert 'keystore' in result
+
+        keystore: Keystore = result['keystore']
+        keystore_path = os.path.join(temp_dir, f'{keystore.identity.id}.json')
+        assert os.path.isfile(keystore_path)
+
+    except CLIRuntimeError:
+        assert False
+
+    # list all identities
+    try:
+        args = {
+            'keystore': temp_dir,
+        }
+
+        cmd = IdentityList()
+        result = cmd.execute(args)
+        assert result is not None
+        assert 'available' in result
+        assert len(result['available']) == 1
+
+    except CLIRuntimeError:
+        assert False
+
+    # show identity
+    try:
+        args = {
+            'keystore': temp_dir,
+            'keystore-id': keystore.identity.id
+        }
+
+        cmd = IdentityShow()
+        result = cmd.execute(args)
+        assert result is not None
+        assert 'content' in result
+        assert result['content'] is not None
+
+    except CLIRuntimeError:
+        assert False
+
+    # remove the identity
+    try:
+        args = {
+            'keystore': temp_dir,
+            'keystore-id': keystore.identity.id,
+            'password': 'password',
+            'confirm': True
+        }
+
+        cmd = IdentityRemove()
+        result = cmd.execute(args)
+        assert result is None
+
+    except CLIRuntimeError:
+        assert False
+
+    # list all identities
+    try:
+        args = {
+            'keystore': temp_dir,
+        }
+
+        cmd = IdentityList()
+        result = cmd.execute(args)
+        assert result is not None
+        assert 'available' in result
+        assert len(result['available']) == 0
+
+    except CLIRuntimeError:
+        assert False
+
+
+
+def test_cli_identity_network(session_node, temp_dir):
+    """Test CLI identity discovery, publishing, and update workflow."""
+    address = session_node.rest.address()
+
+    # create an identity
+    try:
+        args = {
+            'keystore': temp_dir,
+            'name': 'name',
+            'email': 'email',
+            'password': 'password'
+        }
+
+        cmd = IdentityCreate()
+        result = cmd.execute(args)
+        assert result is not None
+        assert 'keystore' in result
+
+        keystore: Keystore = result['keystore']
+        keystore_path = os.path.join(temp_dir, f'{keystore.identity.id}.json')
+        assert os.path.isfile(keystore_path)
+
+    except CLIRuntimeError:
+        assert False
+
+    # discover all identities known to the node
+    try:
+        args = {
+            'address': f"{address[0]}:{address[1]}"
+        }
+
+        cmd = IdentityDiscover()
+        result = cmd.execute(args)
+        assert result is not None
+        assert keystore.identity.id not in result
+
+    except CLIRuntimeError:
+        assert False
+
+    # publish identity
+    try:
+        args = {
+            'keystore': temp_dir,
+            'keystore-id': keystore.identity.id,
+            'password': 'password',
+            'address': f"{address[0]}:{address[1]}"
+        }
+
+        cmd = IdentityPublish()
+        result = cmd.execute(args)
+        assert result is None
+
+    except CLIRuntimeError:
+        assert False
+
+    # discover all identities known to the node
+    try:
+        args = {
+            'address': f"{address[0]}:{address[1]}"
+        }
+
+        cmd = IdentityDiscover()
+        result = cmd.execute(args)
+        assert result is not None
+        assert keystore.identity.id in result
+
+    except CLIRuntimeError:
+        assert False
+
+    # update identity
+    try:
+        args = {
+            'keystore': temp_dir,
+            'keystore-id': keystore.identity.id,
+            'password': 'password',
+            'name': 'updated_name'
+        }
+
+        cmd = IdentityUpdate()
+        result = cmd.execute(args)
+        assert result is not None
+        assert 'keystore' in result
+
+        keystore: Keystore = result['keystore']
+        assert keystore.identity.name == 'updated_name'
+
+    except CLIRuntimeError:
+        assert False
+
+    # publish identity
+    try:
+        args = {
+            'keystore': temp_dir,
+            'keystore-id': keystore.identity.id,
+            'password': 'password',
+            'address': f"{address[0]}:{address[1]}"
+        }
+
+        cmd = IdentityPublish()
+        result = cmd.execute(args)
+        assert result is None
+
+    except CLIRuntimeError:
+        assert False
+
+    # discover all identities known to the node
+    try:
+        args = {
+            'address': f"{address[0]}:{address[1]}"
+        }
+
+        cmd = IdentityDiscover()
+        result = cmd.execute(args)
+        assert result is not None
+        assert keystore.identity.id in result
+
+        identity: Identity = result[keystore.identity.id]
+        assert identity.name == 'updated_name'
+
+    except CLIRuntimeError:
+        assert False
+
+
+
+def test_cli_credentials(temp_dir):
+    """Test CLI credentials management workflow (list, add, remove)."""
+    # create an identity
+    try:
+        args = {
+            'keystore': temp_dir,
+            'name': 'name',
+            'email': 'email',
+            'password': 'password'
+        }
+
+        cmd = IdentityCreate()
+        result = cmd.execute(args)
+        assert result is not None
+        assert 'keystore' in result
+
+        keystore: Keystore = result['keystore']
+        keystore_path = os.path.join(temp_dir, f'{keystore.identity.id}.json')
+        assert os.path.isfile(keystore_path)
+
+    except CLIRuntimeError:
+        assert False
+
+    # list all credentials
+    try:
+        args = {
+            'keystore': temp_dir,
+            'keystore-id': keystore.identity.id,
+            'password': 'password',
+        }
+
+        cmd = CredentialsList()
+        result = cmd.execute(args)
+        assert result is not None
+        assert 'credentials' in result
+        assert len(result['credentials']) == 0
+
+    except CLIRuntimeError:
+        assert False
+
+    # add credentials
+    try:
+        args = {
+            'keystore': temp_dir,
+            'keystore-id': keystore.identity.id,
+            'password': 'password',
+            'url': 'url',
+            'login': 'login',
+            'personal_access_token': 'personal-access-token'
+        }
+
+        cmd = CredentialsAddGithubCredentials()
+        result = cmd.execute(args)
+        assert result is not None
+        assert 'credentials' in result
+        assert result['credentials'] is not None
+
+    except CLIRuntimeError:
+        assert False
+
+    # list all credentials
+    try:
+        args = {
+            'keystore': temp_dir,
+            'keystore-id': keystore.identity.id,
+            'password': 'password',
+        }
+
+        cmd = CredentialsList()
+        result = cmd.execute(args)
+        assert result is not None
+        assert 'credentials' in result
+        assert len(result['credentials']) == 1
+
+    except CLIRuntimeError:
+        assert False
+
+    # remove credentials
+    try:
+        args = {
+            'keystore': temp_dir,
+            'keystore-id': keystore.identity.id,
+            'password': 'password',
+            'credential': 'github:url',
+            'confirm': True
+        }
+
+        cmd = CredentialsRemove()
+        result = cmd.execute(args)
+        assert result is not None
+        assert 'removed' in result
+        assert len(result['removed']) == 1
+
+    except CLIRuntimeError:
+        assert False
+
+    # list all credentials
+    try:
+        args = {
+            'keystore': temp_dir,
+            'keystore-id': keystore.identity.id,
+            'password': 'password',
+        }
+
+        cmd = CredentialsList()
+        result = cmd.execute(args)
+        assert result is not None
+        assert 'credentials' in result
+        assert len(result['credentials']) == 0
+
+    except CLIRuntimeError:
+        assert False
+
+
