+++ simaas/tests/test_dor.py
@@ -1,17 +1,4 @@
-"""Integration tests for the Data Object Repository (DOR) service.
-
-This module tests DOR functionality including:
-- Search and statistics operations
-- Data object creation with various options
-- Data object removal and metadata operations
-- Content retrieval with access control
-- Provenance tracking
-- Access grant/revoke operations
-- Ownership transfer
-- Tag management
-- Content encryption
-- Touch/last accessed tracking
-"""
+"""Integration tests for the Data Object Repository (DOR) service."""
 
 import json
 import random
@@ -38,31 +25,13 @@ logger = Logging.get(__name__)
 
 @pytest.fixture()
 def unknown_user(extra_keystores):
-    """Provide an unknown user keystore (not registered with node).
-
-    Args:
-        extra_keystores: List of available keystores.
-
-    Returns:
-        A Keystore for an unknown user.
-    """
+    """Provide an unknown user keystore (not registered with node)."""
     return extra_keystores[2]
 
 
 @pytest.fixture()
 def known_users(extra_keystores, node_db_proxy):
-    """Provide keystores for users known to the node.
-
-    Registers two users with the node database so they can
-    perform authorized operations.
-
-    Args:
-        extra_keystores: List of available keystores.
-        node_db_proxy: Proxy to node database for identity registration.
-
-    Returns:
-        List of two Keystore instances for known users.
-    """
+    """Provide keystores for users known to the node."""
     keystores = [extra_keystores[0], extra_keystores[1]]
     for keystore in keystores:
         node_db_proxy.update_identity(keystore.identity)
@@ -71,11 +40,7 @@ def known_users(extra_keystores, node_db_proxy):
 
 @pytest.fixture()
 def random_content():
-    """Create a temporary JSON file with random content for testing.
-
-    Yields:
-        Path to a temporary JSON file containing a random integer value.
-    """
+    """Create a temporary JSON file with random content for testing."""
     with tempfile.TemporaryDirectory() as tmpdir:
         content_path = os.path.join(tmpdir, 'test.json')
         with open(content_path, 'w') as f:
@@ -87,14 +52,7 @@ def random_content():
 
 @pytest.fixture()
 def receiver(test_context):
-    """Create a receiver node for data transfer testing.
-
-    Args:
-        test_context: Test context providing node creation utilities.
-
-    Returns:
-        A Node instance configured as a receiver.
-    """
+    """Create a receiver node for data transfer testing."""
     keystore = test_context.create_keystores(1)[0]
     node = test_context.get_node(keystore, enable_rest=True)
     return node
@@ -106,49 +64,21 @@ def receiver(test_context):
 
 @pytest.mark.integration
 def test_search(dor_proxy):
-    """Test DOR search operation.
-
-    Verifies that:
-    - Search operation returns results without error
-    - Result is a list (may be empty)
-
-    Backend: N/A (DOR only)
-    Duration: ~1 second
-    Requirements: None
-    """
+    """Test DOR search operation."""
     result = dor_proxy.search()
     assert (result is not None)
 
 
 @pytest.mark.integration
 def test_statistics(dor_proxy):
-    """Test DOR statistics operation.
-
-    Verifies that:
-    - Statistics operation returns results without error
-    - Result contains repository statistics
-
-    Backend: N/A (DOR only)
-    Duration: ~1 second
-    Requirements: None
-    """
+    """Test DOR statistics operation."""
     result = dor_proxy.statistics()
     assert (result is not None)
 
 
 @pytest.mark.integration
 def test_add_c_multiple_creators(session_keystore, known_users, dor_proxy, random_content):
-    """Test adding data object with multiple creators.
-
-    Verifies that:
-    - Data object can be created with multiple creator identities
-    - Creator list is correctly recorded in metadata
-    - Data object without explicit creators defaults to owner as creator
-
-    Backend: N/A (DOR only)
-    Duration: ~2 seconds
-    Requirements: None
-    """
+    """Test adding data object with multiple creators."""
     owner = session_keystore.identity
     c0 = known_users[0]
     c1 = known_users[1]
@@ -169,17 +99,7 @@ def test_add_c_multiple_creators(session_keystore, known_users, dor_proxy, rando
 
 @pytest.mark.integration
 def test_add_c_license(session_keystore, dor_proxy, random_content):
-    """Test adding data object with license information.
-
-    Verifies that:
-    - Data object can be created with license flags
-    - License BY flag is correctly recorded
-    - Other license flags (SA, NC, ND) default to false
-
-    Backend: N/A (DOR only)
-    Duration: ~1 second
-    Requirements: None
-    """
+    """Test adding data object with license information."""
     owner = session_keystore.identity
 
     result = dor_proxy.add_data_object(random_content, owner, False, False, 'JSON', 'json', license_by=True)
@@ -192,17 +112,7 @@ def test_add_c_license(session_keystore, dor_proxy, random_content):
 
 @pytest.mark.integration
 def test_add_c(session_keystore, dor_proxy, unknown_user, random_content):
-    """Test adding data object with identity validation.
-
-    Verifies that:
-    - Adding with unknown owner fails with 'Identity not found' error
-    - Adding with unknown creator fails with 'Identity not found' error
-    - Adding with valid identities succeeds
-
-    Backend: N/A (DOR only)
-    Duration: ~2 seconds
-    Requirements: None
-    """
+    """Test adding data object with identity validation."""
     owner = session_keystore.identity
 
     # unknown owner
@@ -222,17 +132,7 @@ def test_add_c(session_keystore, dor_proxy, unknown_user, random_content):
 @pytest.mark.integration
 @pytest.mark.slow
 def test_add_c_large(test_context, session_keystore, dor_proxy):
-    """Test adding large data objects and measure upload performance.
-
-    Verifies that:
-    - Large files (252 MB) can be uploaded successfully
-    - Content hash is correctly computed for large files
-    - Upload performance is measured and logged
-
-    Backend: N/A (DOR only)
-    Duration: ~60 seconds
-    Requirements: None
-    """
+    """Test adding large data objects and measure upload performance."""
     # this is necessary to avoid getting thousands of 'Calling on_part_data with data' messages...
     mp_logger = logging.getLogger('python_multipart.multipart')
     mp_logger.setLevel(logging.INFO)
@@ -278,18 +178,7 @@ def test_add_c_large(test_context, session_keystore, dor_proxy):
 
 @pytest.mark.integration
 def test_remove(dor_proxy, random_content, known_users):
-    """Test data object removal with authorization.
-
-    Verifies that:
-    - Deleting non-existent object fails with 'Data object not found' error
-    - Deleting with wrong authority fails with 'user is not the data object owner'
-    - Deleting with correct authority succeeds
-    - Deleted object is no longer retrievable
-
-    Backend: N/A (DOR only)
-    Duration: ~2 seconds
-    Requirements: None
-    """
+    """Test data object removal with authorization."""
     c0 = known_users[0]
     c1 = known_users[1]
 
@@ -317,17 +206,7 @@ def test_remove(dor_proxy, random_content, known_users):
 
 @pytest.mark.integration
 def test_get_meta(session_keystore, dor_proxy, random_content):
-    """Test getting data object metadata.
-
-    Verifies that:
-    - Getting metadata for invalid object returns None
-    - Getting metadata for valid object returns correct information
-    - Object ID matches in returned metadata
-
-    Backend: N/A (DOR only)
-    Duration: ~2 seconds
-    Requirements: None
-    """
+    """Test getting data object metadata."""
     owner = session_keystore.identity
 
     result = dor_proxy.add_data_object(random_content, owner, False, False, 'JSON', 'json')
@@ -344,18 +223,7 @@ def test_get_meta(session_keystore, dor_proxy, random_content):
 
 @pytest.mark.integration
 def test_get_content(test_context, session_keystore, dor_proxy, random_content, unknown_user, known_users):
-    """Test getting data object content with access control.
-
-    Verifies that:
-    - Getting content for non-existent object fails with appropriate error
-    - Getting content with unknown authority fails with 'unknown identity'
-    - Getting content without access permission fails with 'user has no access'
-    - Getting content with correct authority succeeds and downloads file
-
-    Backend: N/A (DOR only)
-    Duration: ~3 seconds
-    Requirements: None
-    """
+    """Test getting data object content with access control."""
     owner = session_keystore.identity
 
     result = dor_proxy.add_data_object(random_content, owner, True, False, 'JSON', 'json')
@@ -386,18 +254,7 @@ def test_get_content(test_context, session_keystore, dor_proxy, random_content, 
 
 @pytest.mark.integration
 def test_get_provenance(test_context, session_keystore, dor_proxy):
-    """Test data object provenance tracking.
-
-    Verifies that:
-    - Provenance can be retrieved for data objects
-    - Data nodes are correctly tracked in provenance graph
-    - Recipe information creates proper provenance steps
-    - Processor and input/output hashes are correctly recorded
-
-    Backend: N/A (DOR only)
-    Duration: ~5 seconds
-    Requirements: None
-    """
+    """Test data object provenance tracking."""
     processor = {
         'repository': 'github.com/source',
         'commit_id': '34534ab',
@@ -480,21 +337,7 @@ def test_get_provenance(test_context, session_keystore, dor_proxy):
 
 @pytest.mark.integration
 def test_grant_revoke_access(session_keystore, dor_proxy, random_content, known_users):
-    """Test granting and revoking access to data objects.
-
-    Verifies that:
-    - Initial access list contains only owner
-    - Non-owner cannot grant access
-    - Owner can grant access to other users
-    - Granting access to user who already has access is idempotent
-    - Non-owner cannot revoke access
-    - Owner can revoke access from users
-    - Revoking access from user without access is idempotent
-
-    Backend: N/A (DOR only)
-    Duration: ~3 seconds
-    Requirements: None
-    """
+    """Test granting and revoking access to data objects."""
     owner = session_keystore
 
     result = dor_proxy.add_data_object(random_content, owner.identity, False, False, 'JSON', 'json')
@@ -550,18 +393,7 @@ def test_grant_revoke_access(session_keystore, dor_proxy, random_content, known_
 
 @pytest.mark.integration
 def test_transfer_ownership(session_keystore, dor_proxy, random_content, known_users, unknown_user):
-    """Test transferring data object ownership.
-
-    Verifies that:
-    - Non-owner cannot transfer ownership
-    - Cannot transfer ownership to unknown user
-    - Owner can transfer ownership to known user
-    - New owner is correctly reflected in metadata
-
-    Backend: N/A (DOR only)
-    Duration: ~2 seconds
-    Requirements: None
-    """
+    """Test transferring data object ownership."""
     owner = session_keystore
 
     meta = dor_proxy.add_data_object(random_content, owner.identity, False, False, 'JSON', 'json')
@@ -591,22 +423,7 @@ def test_transfer_ownership(session_keystore, dor_proxy, random_content, known_u
 
 @pytest.mark.integration
 def test_update_remove_tags(session_keystore, dor_proxy, random_content, known_users):
-    """Test updating and removing tags on data objects.
-
-    Verifies that:
-    - Initial object has no tags
-    - Non-owner cannot update tags
-    - Owner can add tags
-    - Owner can update existing tags
-    - Non-owner cannot remove tags
-    - Removing non-existent tag is safe (no error)
-    - Owner can remove existing tags
-    - Complex tag values (nested objects) are supported
-
-    Backend: N/A (DOR only)
-    Duration: ~3 seconds
-    Requirements: None
-    """
+    """Test updating and removing tags on data objects."""
     owner = session_keystore
 
     result = dor_proxy.add_data_object(random_content, owner.identity, False, False, 'JSON', 'json')
@@ -660,18 +477,7 @@ def test_update_remove_tags(session_keystore, dor_proxy, random_content, known_u
 
 @pytest.mark.integration
 def test_content_encryption(test_context, known_users, dor_proxy):
-    """Test content encryption and decryption workflow.
-
-    Verifies that:
-    - Encrypted content can be uploaded as data object
-    - Ownership can be transferred for encrypted objects
-    - Content key can be used to decrypt content after transfer
-    - Decrypted content matches original plaintext
-
-    Backend: N/A (DOR only)
-    Duration: ~3 seconds
-    Requirements: None
-    """
+    """Test content encryption and decryption workflow."""
     # create content for the data object and encrypt it
     content_plain = "my little secret..."
     content_enc, content_key = symmetric_encrypt(content_plain.encode('utf-8'))
@@ -701,18 +507,7 @@ def test_content_encryption(test_context, known_users, dor_proxy):
 
 @pytest.mark.integration
 def test_search_by_content_hashes(test_context, known_users, dor_proxy):
-    """Test searching for data objects by content hashes.
-
-    Verifies that:
-    - Multiple data objects can be created with unique content
-    - Search by content hashes returns correct objects
-    - Multiple hashes can be searched simultaneously
-    - Single hash search returns single result
-
-    Backend: N/A (DOR only)
-    Duration: ~5 seconds
-    Requirements: None
-    """
+    """Test searching for data objects by content hashes."""
     owner = known_users[0]
 
     # create data objects
@@ -751,20 +546,7 @@ def test_search_by_content_hashes(test_context, known_users, dor_proxy):
 
 @pytest.mark.integration
 def test_touch_data_object(test_context, known_users, dor_proxy, random_content):
-    """Test last_accessed timestamp updates for data objects.
-
-    Verifies that:
-    - Getting metadata does NOT update last_accessed
-    - Getting content DOES update last_accessed
-    - Getting provenance does NOT update last_accessed
-    - Access operations (grant/revoke) DO update last_accessed
-    - Ownership transfer DOES update last_accessed
-    - Tag operations (update/remove) DO update last_accessed
-
-    Backend: N/A (DOR only)
-    Duration: ~5 seconds
-    Requirements: None
-    """
+    """Test last_accessed timestamp updates for data objects."""
     owner = known_users[0]
 
     # create data object
