+++ simaas/tests/test_unit_core.py
@@ -1,16 +1,4 @@
-"""Unit tests for core simaas functionality.
-
-This module tests:
-- EC key pair serialization and signing
-- RSA key pair serialization, signing, and encryption
-- Keystore creation, loading, and updating
-- Logging configuration and features
-- Exception handling
-
-Backend: None (unit tests)
-Duration: ~30 seconds
-Requirements: None
-"""
+"""Unit tests for core simaas functionality."""
 import logging
 import os
 
@@ -48,15 +36,7 @@ def logging():
 
 
 def test_ec_serialisation(temp_directory, ec_keypair):
-    """
-    Test EC key pair serialization and deserialization.
-
-    Verifies that:
-    - Public keys can be written to files and read back
-    - Private keys can be written to files and read back
-    - Keys can be serialized to/from bytes and strings
-    - All deserialization methods produce keys with matching IIDs
-    """
+    """Test EC key pair serialization and deserialization."""
     password = 'test'
 
     pubkey_path = os.path.join(temp_directory, 'pubkey.pem')
@@ -108,14 +88,7 @@ def test_ec_serialisation(temp_directory, ec_keypair):
 
 
 def test_ec_signing(ec_keypair):
-    """
-    Test EC key pair signing and verification.
-
-    Verifies that:
-    - Messages can be signed with private key
-    - Signatures can be verified with public key
-    - Wrong messages fail verification
-    """
+    """Test EC key pair signing and verification."""
     message0 = 'test0'.encode('utf-8')
     message1 = 'test1'.encode('utf-8')
 
@@ -125,15 +98,7 @@ def test_ec_signing(ec_keypair):
 
 
 def test_rsa_serialisation(temp_directory, rsa_keypair):
-    """
-    Test RSA key pair serialization and deserialization.
-
-    Verifies that:
-    - Public keys can be written to files and read back
-    - Private keys can be written to files and read back
-    - Keys can be serialized to/from bytes and strings
-    - All deserialization methods produce keys with matching IIDs
-    """
+    """Test RSA key pair serialization and deserialization."""
     password = 'test'
 
     pubkey_path = os.path.join(temp_directory, 'pubkey.pem')
@@ -185,14 +150,7 @@ def test_rsa_serialisation(temp_directory, rsa_keypair):
 
 
 def test_rsa_signing(rsa_keypair):
-    """
-    Test RSA key pair signing and verification.
-
-    Verifies that:
-    - Messages can be signed with private key
-    - Signatures can be verified with public key
-    - Wrong messages fail verification
-    """
+    """Test RSA key pair signing and verification."""
     message0 = 'test0'.encode('utf-8')
     message1 = 'test1'.encode('utf-8')
 
@@ -202,14 +160,7 @@ def test_rsa_signing(rsa_keypair):
 
 
 def test_rsa_encryption(rsa_keypair):
-    """
-    Test RSA encryption and decryption.
-
-    Verifies that:
-    - Plaintext can be encrypted with public key
-    - Ciphertext can be decrypted with private key
-    - Decrypted text matches original
-    """
+    """Test RSA encryption and decryption."""
     plaintext = "test"
 
     encrypted = rsa_keypair.encrypt(plaintext.encode('utf-8'))
@@ -219,15 +170,7 @@ def test_rsa_encryption(rsa_keypair):
 
 
 def test_create_and_load(temp_directory):
-    """
-    Test keystore creation and loading.
-
-    Verifies that:
-    - New keystore can be created with name and email
-    - Keystore is saved to file
-    - Keystore can be loaded from file
-    - Loaded keystore has correct identity
-    """
+    """Test keystore creation and loading."""
     keystore = Keystore.new('name', 'email', path=temp_directory, password='password')
     assert(keystore is not None)
     assert(keystore.identity.name == 'name')
@@ -247,15 +190,7 @@ def test_create_and_load(temp_directory):
 
 
 def test_update(temp_directory):
-    """
-    Test keystore profile update.
-
-    Verifies that:
-    - Profile can be updated with new name and email
-    - Updated identity has valid signature
-    - Changes persist after reload
-    - Nonce is incremented
-    """
+    """Test keystore profile update."""
     keystore = Keystore.new('name', 'email', path=temp_directory, password='password')
     keystore_id = keystore.identity.id
     assert(keystore.identity.name == 'name')
@@ -311,14 +246,7 @@ def test_add_get_object_key(temp_directory):
 
 
 def test_add_credentials(temp_directory):
-    """
-    Test keystore credentials management.
-
-    Verifies that:
-    - GitHub credentials can be stored and retrieved
-    - SSH credentials can be stored and retrieved
-    - Credentials persist after sync and reload
-    """
+    """Test keystore credentials management."""
     url = 'https://github.com/sec-digital-twin-lab/saas-middleware'
     login = 'johndoe'
     personal_access_token = 'token'
@@ -352,14 +280,7 @@ def test_add_credentials(temp_directory):
 
 
 def test_defaults(logging):
-    """
-    Test default logging configuration.
-
-    Verifies that:
-    - Logging can be initialized with defaults
-    - INFO messages are visible
-    - DEBUG messages are not visible (default level)
-    """
+    """Test default logging configuration."""
     Logging.initialise()
     logger = Logging.get('test')
 
@@ -368,14 +289,7 @@ def test_defaults(logging):
 
 
 def test_log_to_separate_file(logging, temp_directory):
-    """
-    Test logging to separate files.
-
-    Verifies that:
-    - Default logger writes to default log file
-    - Custom logger writes to both default and custom files
-    - Correct number of lines in each file
-    """
+    """Test logging to separate files."""
     default_log_path = os.path.join(temp_directory, 'log.default')
     custom_log_path = os.path.join(temp_directory, 'log.custom')
 
@@ -434,13 +348,7 @@ def test_rollover(logging, temp_directory):
 
 
 def test_json_incompatible_exception():
-    """
-    Test exception handling with non-JSON-compatible details.
-
-    Verifies that:
-    - SaaSRuntimeException handles non-serializable objects
-    - All detail values are converted to strings
-    """
+    """Test exception handling with non-JSON-compatible details."""
     class SomeClass:
         pass
 
@@ -463,17 +371,7 @@ def test_json_incompatible_exception():
 
 
 def test_logging_aws_cloudwatch_integration(logging):
-    """
-    Test AWS CloudWatch logging integration.
-
-    Verifies that:
-    - Logging can be initialized with AWS CloudWatch enabled
-    - Messages can be logged without errors
-
-    Backend: None (unit test)
-    Duration: <5 seconds
-    Requirements: AWS credentials in .aws directory (optional - test passes without)
-    """
+    """Test AWS CloudWatch logging integration."""
     # Before running this test, ensure that [default] in credentials file is available in .aws directory
     # Initialize logging with AWS CloudWatch enabled
     Logging.initialise(log_to_aws=True)
